feat: create the search bar
